Extract shared test infrastructure into Engraved.TestUtils
Engraved.Persistence.Mongo.Tests previously played a dual role: it
contained the persistence tests AND the shared test infrastructure
(TestMongoRepository, Util/EphemeralMongo runner, MockCurrentUserService,
TestIds, EngravedTestContext), which forced Core.Tests and Api.Tests to
reference a test project.

The infrastructure now lives in Engraved.TestUtils, a plain class
library without the test SDK, referenced by all three test projects.

MongoRunnerTeardown keeps the linked-file pattern (NUnit only discovers
SetUpFixtures in the assembly under test); the file is hosted in
Engraved.TestUtils but excluded from its own compilation, and all three
test projects link it - including Engraved.Persistence.Mongo.Tests,
which previously compiled it directly.
