+++ api/Engraved.Api.Tests/Engraved.Api.Tests.csproj
@@ -8,7 +8,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Engraved.Api\Engraved.Api.csproj"/>
-        <ProjectReference Include="..\Engraved.Persistence.Mongo.Tests\Engraved.Persistence.Mongo.Tests.csproj"/>
+        <ProjectReference Include="..\Engraved.TestUtils\Engraved.TestUtils.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
@@ -26,7 +26,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Compile Include="..\Engraved.Persistence.Mongo.Tests\Source\MongoRunnerTeardown.cs">
+      <Compile Include="..\Engraved.TestUtils\Source\MongoRunnerTeardown.cs">
         <Link>Source\MongoRunnerTeardown.cs</Link>
       </Compile>
     </ItemGroup>
