+++ api/Engraved.Api.Tests/Source/Authentication/RefreshHandlerShould.cs
@@ -3,7 +3,7 @@ using Engraved.Api.Authentication;
 using Engraved.Api.Settings;
 using Engraved.Core.Application;
 using Engraved.Core.Application.Persistence;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using Engraved.Core.Domain.Users;
 using FluentAssertions;
 using NUnit.Framework;
