+++ api/Engraved.Core.Tests/Source/Application/Commands/Entries/Delete/DeleteEntryCommandExecutorShould.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Permissions;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
