+++ api/Engraved.Core.Tests/Source/Application/Commands/Entries/Move/MoveEntryCommandExecutorShould.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
