+++ api/Engraved.Core.Tests/Source/Application/Commands/Entries/Upsert/UpsertCounterEntryCommandExecutorShould.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Engraved.Core.Application.Commands.Entries.Upsert.Counter;
 using Engraved.Core.Application.Commands.Entries.Upsert.Gauge;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using FluentAssertions;
