+++ api/Engraved.Core.Tests/Source/Application/Commands/Entries/Upsert/UpsertTimerEntryCommandExecutorShould.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Engraved.Core.Application.Commands.Entries.Upsert.Timer;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using FluentAssertions;
