+++ api/Engraved.Core.Tests/Source/Application/Commands/Users/AddJournalToFavorites/AddJournalToFavoritesCommandExecutorShould.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Engraved.Core.Domain.Users;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
