+++ api/Engraved.Core.Tests/Source/Application/Commands/Users/UpdateTags/UpdateUserTagsCommandExecutorShould.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Engraved.Core.Domain.Users;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
