+++ api/Engraved.Core.Tests/Source/Application/Queries/Entries/GetAll/SearchEntriesQueryExecutorShould.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Users;
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
