+++ api/Engraved.Persistence.Mongo.Tests/Source/JournalAccessFilter_Matches_JournalAccessPolicy_Should.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Permissions;
 using Engraved.Persistence.Mongo.DocumentTypes.Journals;
+using Engraved.TestUtils;
 using FluentAssertions;
 using MongoDB.Driver;
 using NUnit.Framework;
