+++ api/Engraved.Persistence.Mongo.Tests/Source/MongoRepositoryBaseShould.cs
@@ -5,6 +5,7 @@ using Engraved.Core.Application.Persistence;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Users;
+using Engraved.TestUtils;
 using FluentAssertions;
 using Newtonsoft.Json;
 using NUnit.Framework;
