+++ api/Engraved.Persistence.Mongo.Tests/Source/MongoRepositoryBase_GetEntriesForJournal_Should.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Engraved.Core.Application.Persistence;
 using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
