+++ api/Engraved.Persistence.Mongo.Tests/Source/UserRestrictedMongoRepositoryShould.cs
@@ -6,6 +6,7 @@ using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Users;
 using Engraved.Persistence.Mongo.DocumentTypes.Entries;
 using Engraved.Persistence.Mongo.DocumentTypes.Journals;
+using Engraved.TestUtils;
 using FluentAssertions;
 using MongoDB.Driver;
 using NUnit.Framework;
