+++ api/Engraved.Persistence.Mongo.Tests/Source/UserRestrictedMongoRepository_Permissions_Should.cs
@@ -6,6 +6,7 @@ using Engraved.Core.Domain.Entries;
 using Engraved.Core.Domain.Journals;
 using Engraved.Core.Domain.Permissions;
 using Engraved.Core.Domain.Users;
+using Engraved.TestUtils;
 using FluentAssertions;
 using NUnit.Framework;
