+++ api/Engraved.TestUtils/Source/MongoRunnerTeardown.cs
@@ -1,4 +1,4 @@
-using Engraved.Persistence.Mongo.Tests;
+using Engraved.TestUtils;
 using NUnit.Framework;
 
 // Assembly-level teardown: NUnit runs this once after every test in the assembly has finished.
